+++ verticapy/machine_learning/memmodel/memmodel.py
@@ -26,7 +26,6 @@ from typing import Union, Literal
 
 # VerticaPy Modules
 from verticapy.utils._decorators import save_verticapy_logs
-from verticapy.utils._toolbox import *
 from verticapy.errors import *
 from verticapy.sql._utils._format import clean_query
