+++ verticapy/machine_learning/vertica/decomposition.py
@@ -25,7 +25,6 @@ from verticapy.utils._decorators import (
     check_minimum_version,
 )
 from verticapy.utilities import *
-from verticapy.utils._toolbox import *
 from verticapy.learn.vmodel import *
 
 # Standard Module
