+++ verticapy/machine_learning/vertica/pipeline.py
@@ -26,7 +26,6 @@ from verticapy.utils._decorators import (
 )
 from verticapy import vDataFrame
 from verticapy.utilities import *
-from verticapy.utils._toolbox import *
 from verticapy.errors import *
 from verticapy.learn.vmodel import *
