+++ verticapy/machine_learning/vertica/tools.py
@@ -24,7 +24,6 @@ from verticapy.utils._decorators import (
     save_verticapy_logs,
     check_minimum_version,
 )
-from verticapy.utils._toolbox import *
 from verticapy.sql.read import _executeSQL
 from verticapy.utilities import *
 from verticapy.sql._utils._format import quote_ident, schema_relation
