+++ verticapy/plotting/_matplotlib/bar.py
@@ -23,7 +23,7 @@ import matplotlib.pyplot as plt
 
 # VerticaPy Modules
 from verticapy.utilities import *
-from verticapy.utils._toolbox import updated_dict
+from verticapy.plotting._matplotlib.core import updated_dict
 from verticapy._config._notebook import ISNOTEBOOK
 from verticapy.errors import ParameterError
 from verticapy.plotting._matplotlib.core import compute_plot_variables
