+++ verticapy/plotting/_matplotlib/pie.py
@@ -24,7 +24,7 @@ import numpy as np
 # VerticaPy Modules
 from verticapy.utilities import *
 from verticapy.sql._utils._format import quote_ident
-from verticapy.utils._toolbox import updated_dict
+from verticapy.plotting._matplotlib.core import updated_dict
 from verticapy._config._notebook import ISNOTEBOOK
 from verticapy.sql.read import _executeSQL
 from verticapy.errors import ParameterError
