+++ verticapy/sql/parsers/csv.py
@@ -25,7 +25,6 @@ import os, warnings
 # VerticaPy Modules
 import verticapy as vp
 from verticapy.utils._decorators import save_verticapy_logs
-from verticapy.utils._toolbox import *
 from verticapy.utils._gen import gen_tmp_name
 from verticapy.sql.read import _executeSQL
 from verticapy.errors import ExtensionError, ParameterError, MissingRelation
