+++ verticapy/stats/tools.py
@@ -34,7 +34,6 @@ from verticapy.utils._decorators import (
     check_minimum_version,
 )
 from verticapy.utilities import *
-from verticapy.utils._toolbox import *
 from verticapy.utils._gen import gen_tmp_name
 from verticapy.learn.linear_model import LinearRegression
 from verticapy import vDataFrame
